fix the bug that images were not not deleting from cloudinary
