+++ Server/Router/UserRouter.js
@@ -3,9 +3,7 @@ const router = express.Router();
 const multer = require("multer");
 
 const { AddUser, EditUser, CheckUser , GetUser} = require("../Controller/SPController");
-const {
-  user_rating_and_feedback,
-} = require("../Controller/WorkCompletedController");
+const {user_rating_and_feedback} = require("../Controller/WorkCompletedController");
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
